+++ src/app/website/documentation/javascript-sdk/getting-started.view.ts
@@ -1,20 +1,27 @@
 import { Component, ElementRef, ChangeDetectorRef } from "@angular/core";
-import { CORE_DIRECTIVES } from "@angular/common";
+import { Location } from "@angular/common";
 import { ROUTER_DIRECTIVES, Router } from "@angular/router-deprecated";
 import { Title } from "@angular/platform-browser";
 
-import SidebarComponent from "./../../sidebar/SidebarComponent";
+
+import HighlightDirective from "carbon-panel/directives/highlight.directive";
+import { TabsComponent } from "carbon-panel/semantic/tabs.component";
+import { TabComponent } from "carbon-panel/semantic/tab.component";
+
+import SidebarComponent from "./../sidebar/SidebarComponent";
+
 
 import highlight from "highlight.js";
 import "highlight.js/styles/tomorrow-night.css!";
 
-import template from "./template.html!";
-
+import template from "./getting-started.view.html!";
+import style from "./getting-started.view.css!text";
 
 @Component( {
 	selector: "getting-started",
 	template: template,
-	directives: [ CORE_DIRECTIVES, ROUTER_DIRECTIVES, SidebarComponent ],
+	styles: [ style ],
+	directives: [ ROUTER_DIRECTIVES, SidebarComponent, HighlightDirective, TabsComponent, TabComponent ],
 	providers: [ Title ],
 } )
 export default class GettingStartedView {
@@ -22,18 +29,17 @@ export default class GettingStartedView {
 	element:ElementRef;
 	$element:JQuery;
 	title:Title;
-
 	contentReady:boolean = false;
 
+	private location:Location;
 	private changeDetector:ChangeDetectorRef;
 
-	constructor( router:Router, element:ElementRef, title:Title, changeDetector:ChangeDetectorRef ) {
+	constructor( element:ElementRef, title:Title, router:Router, location:Location, changeDetector:ChangeDetectorRef ) {
 		this.element = element;
-		this.router = router;
 		this.title = title;
-
+		this.router = router;
+		this.location = location;
 		this.changeDetector = changeDetector;
-
 	}
 
 	ngAfterViewInit():void {
@@ -41,7 +47,7 @@ export default class GettingStartedView {
 		this.initializeAccordions();
 		this.initializeTabs();
 		this.highlightCode();
-
+		this.$element.find( ".sectionlink a[href]" ).on( "click", this.scrollTo );
 		this.initializeSidebar();
 	}
 
@@ -72,4 +78,18 @@ export default class GettingStartedView {
 		}, 0 );
 	}
 
+	scrollTo( event:any ):boolean {
+		let id:string = $( event.currentTarget ).attr( "href" ).replace( "#", "" );
+		let $element:JQuery = $( "#" + id );
+		let position:number = $element.offset().top - 100;
+
+		$( "html, body" ).animate( {
+			scrollTop: position
+		}, 500 );
+		location.hash = "#" + id;
+		event.stopImmediatePropagation();
+		event.preventDefault();
+
+		return false;
+	}
 }